+++ test/mitmproxy/addons/test_proxyserver.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 import asyncio
-import socket
 import ssl
 from collections.abc import AsyncGenerator
 from collections.abc import Callable
@@ -266,22 +265,18 @@ async def test_shutdown_err(caplog_async) -> None:
         await _wait_for_connection_closes(ps)
 
 
-class DummyResolver:
-    async def dns_request(self, flow: dns.DNSFlow) -> None:
-        flow.response = await dns_resolver.resolve_message(flow.request, self)
-
-    async def getaddrinfo(self, host: str, port: int, *, family: int, type: int):
-        if family == socket.AF_INET and host == "dns.google":
-            return [(socket.AF_INET, type, None, None, ("8.8.8.8", port))]
-        e = socket.gaierror()
-        e.errno = socket.EAI_NONAME
-        raise e
+async def lookup_ipv4():
+    return await asyncio.sleep(0, ["8.8.8.8"])
 
 
-async def test_dns(caplog_async) -> None:
+async def test_dns(caplog_async, monkeypatch) -> None:
+    monkeypatch.setattr(
+        mitmproxy_rs.DnsResolver, "lookup_ipv4", lambda _, __: lookup_ipv4()
+    )
+
     caplog_async.set_level("INFO")
     ps = Proxyserver()
-    with taddons.context(ps, DummyResolver()) as tctx:
+    with taddons.context(ps, dns_resolver.DnsResolver()) as tctx:
         tctx.configure(
             ps,
             mode=["dns@127.0.0.1:0"],