+++ CHANGELOG.md
@@ -31,6 +31,8 @@
   ([#7001](https://github.com/mitmproxy/mitmproxy/pull/7001), @errorxyz)
 * Add Host header to CONNECT requests.
   ([#7021](https://github.com/mitmproxy/mitmproxy/pull/7021), @petsneakers)
+* Support all query types in DNS mode
+  ([#6975](https://github.com/mitmproxy/mitmproxy/pull/6975), @errorxyz)
 
 ## 12 June 2024: mitmproxy 10.3.1